model_service/requirements.txt
flask
kafka-python
pandas
scikit-learn
joblib
sqlalchemy
psycopg2-binary
prometheus_client
mlflow
